Add comprehensive unit tests for Thread and parallel_for_each_optimized
- Implemented tests for the Thread class covering basic functionality, exception handling, stop token signaling, and thread management methods.
- Added tests for parallel_for_each_optimized to validate its behavior with various data types, concurrency scenarios, and edge cases.
- Included tests for OptimizedTask, CacheAligned, SpinLock, RWSpinLock, and SPSCQueue to ensure thread safety and performance.
- Ensured coverage for both normal and exceptional cases, including empty ranges and single-element scenarios.
- Verified correct handling of exceptions in tasks and the behavior of synchronization primitives under concurrent access.

+++ atom/algorithm/pathfinding.cpp
@@ -406,8 +406,8 @@ std::optional<std::vector<Point>> PathFinder::findJPSPath(const GridMap& map,
 
             f32 tentativeG = gScore[current];
 
-            f32 dx = jumpPoint->x - current.x;
-            f32 dy = jumpPoint->y - current.y;
+            f32 dx = static_cast<f32>(jumpPoint->x - current.x);
+            f32 dy = static_cast<f32>(jumpPoint->y - current.y);
             f32 dist = std::sqrt(dx * dx + dy * dy);
 
             tentativeG += dist * 1.0f;
